Reorganized file structure, putting components and their interfaces into folders.
